requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.3.0
tqdm>=4.65.0

# Data science
scikit-learn>=1.2.0
pandas>=1.5.0

# Web demo
gradio>=4.0.0

# API and deployment
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# MLOps
mlflow>=2.5.0
tensorboard>=2.11.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0
mkdocstrings[python]>=0.20.0